requirements.txt
numpy
matplotlib
cvxpy
jaxlib
matplotlib
ipyopt